+++ mindspore/lite/examples/train_lenet_java/model/prepare_model.sh
@@ -1,17 +1,19 @@
 #!/bin/bash
 
-echo "============Exporting=========="
-if [ -n "$1" ]; then
-  DOCKER_IMG=$1
-  docker run -w $PWD --runtime=nvidia -v /home/$USER:/home/$USER --privileged=true ${DOCKER_IMG} /bin/bash -c "PYTHONPATH=../../../../../model_zoo/official/cv/lenet/src python lenet_export.py; chmod 444 lenet_tod.mindir; rm -rf __pycache__"
-else
-  echo "MindSpore docker was not provided, attempting to run locally"
-  PYTHONPATH=../../../../../model_zoo/official/cv/lenet/src python lenet_export.py
-fi
-
-if [ ! -f "$CONVERTER" ]; then
-  echo "converter_lite could not be found in MindSpore build directory nor in system path"
-  exit 1
+if [[ -z ${EXPORT} ]]; then
+  echo "============Exporting=========="
+  if [ -n "$1" ]; then
+    DOCKER_IMG=$1
+    docker run -w $PWD --runtime=nvidia -v /home/$USER:/home/$USER --privileged=true ${DOCKER_IMG} /bin/bash -c "PYTHONPATH=../../../../../model_zoo/official/cv/lenet/src python lenet_export.py; chmod 444 lenet_tod.mindir; rm -rf __pycache__"
+  else
+    echo "MindSpore docker was not provided, attempting to run locally"
+    PYTHONPATH=../../../../../model_zoo/official/cv/lenet/src python lenet_export.py
+  fi
+  
+  if [ ! -f "$CONVERTER" ]; then
+    echo "converter_lite could not be found in MindSpore build directory nor in system path"
+    exit 1
+  fi
 fi
 
 echo "============Converting========="
